Replace address type to Addr
