+++ packages/desmos-std/src/subspaces/query_router.rs
@@ -1,4 +1,4 @@
-use cosmwasm_std::{StdResult};
+use cosmwasm_std::{StdResult, Addr};
 use crate::{
     types::PageRequest,
     subspaces::{
@@ -19,5 +19,5 @@ pub trait SubspacesQuerier {
     fn query_user_groups(&self, subspace_id: u64, pagination: Option<PageRequest>) -> StdResult<QueryUserGroupsResponse>;
     fn query_user_group(&self, subspace_id: u64, group_id : u32) -> StdResult<QueryUserGroupResponse>;
     fn query_user_group_members(&self, subspace_id: u64, group_id : u32, pagination: Option<PageRequest>) -> StdResult<QueryUserGroupMembersResponse>;
-    fn query_user_permissions(&self, subspace_id: u64, user : String) -> StdResult<QueryUserPermissionsResponse>;
+    fn query_user_permissions(&self, subspace_id: u64, user : Addr) -> StdResult<QueryUserPermissionsResponse>;
 }
